Create initial MS fciv.sls ver. 2.05
Created initial MS fciv.sls ver. 2.05. Don't really like this way of installing software. dropping it in c:\windows directory. but salt needs to learn to deal with apps that do not have an installer, until then thsi will have to do. plus the fciv.xe is not a dangerous executable prone to exploitation and I do have a tested and working uninstaller for salt , so if salt tries to remove it , that will succeed and put the machine back to how it was.