Create initial gpu-z.sls ver. 0.8.1
Created initial gpu-z.sls ver. 0.8.1
This installer needs a bit of work. the /S switch will launch it and then the actual installer needs to be laucnhed from it's menu. This will need automating for this to work with salt. But on the flipside the /S for unistall does work.
RFE logged upstream for them to add a properly silent  /S switch to installer.