undo direct commit to do a PR instead
undo direct commit to do a PR instead, sorry I was not used to have commit rights and it committed diorect now. ;-( will switch my workflow to explicitly create a PR from now on. in the past the missing rights forced a PR but now I have to opt for PR instead of commiting direct