Update the installer/uninstaller URLs for the existing nsclient versions, these recently changed and broke this package sls. I'm choosing not to swap the existing versions to salt:// references since that would be a pretty confusing change for people currently using the formula.
However, I've also added the most recent nsclient version with a salt:// reference. It's much better to force the administrator to locally host the installer than to force them to use a link that is likely to break at some indefinite point in the future. For all new versions going forward, I would strongly recommend doing this.

Fixes #586